WIP! Move code to separate files
